+++ docs/examples_source/08_creative_visualizations/plot_particles_fireworks.py
@@ -124,7 +124,8 @@ ax.fill_between([-10, 10], -2, -2, color="oc.blue9", alpha=0.3)
 ax.set_xlim(-10, 10)
 ax.set_ylim(-2, 10)
 ax.set_aspect("equal")
-dm.hide_all_spines(ax)
+for s in ax.spines.values():
+    s.set_visible(False)
 ax.set_facecolor("oc.gray9")
 
 ax.text(