+++ docs/examples_source/08_creative_visualizations/plot_particles_flow_dynamics.py
@@ -122,7 +122,8 @@ for vx, vy in vortices:
 ax.set_xlim(-4, 4)
 ax.set_ylim(-4, 4)
 ax.set_aspect("equal")
-dm.hide_all_spines(ax)
+for s in ax.spines.values():
+    s.set_visible(False)
 ax.set_facecolor("oc.gray1")
 
 ax.text(